[refactor] follow-block 패키지 분리
