+++ src/main/java/com/traveljournal/domain/block/service/BlockService.java
@@ -1,19 +1,16 @@
-package com.traveljournal.domain.member.service;
+package com.traveljournal.domain.block.service;
 
-import com.traveljournal.domain.member.dto.BlockResponse;
-import com.traveljournal.domain.member.entity.Block;
+import com.traveljournal.domain.block.dto.BlockResponse;
+import com.traveljournal.domain.block.entity.Block;
 import com.traveljournal.domain.member.entity.Member;
-import com.traveljournal.domain.member.repository.BlockRepository;
+import com.traveljournal.domain.block.repository.BlockRepository;
+import com.traveljournal.domain.member.service.MemberService;
 import com.traveljournal.global.exception.BlockBadRequestException;
-import com.traveljournal.global.exception.FollowBadRequestException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
