+++ src/main/java/com/traveljournal/domain/follow/entity/RequestStatus.java
@@ -1,4 +1,4 @@
-package com.traveljournal.domain.member.entity;
+package com.traveljournal.domain.follow.entity;
 
 public enum RequestStatus {
     ACCEPTED, REQUESTED, REJECTED
