+++ src/main/java/com/traveljournal/domain/search/service/MemberSearchService.java
@@ -2,7 +2,7 @@ package com.traveljournal.domain.search.service;
 
 import com.traveljournal.domain.member.dto.MemberProfileResponse;
 import com.traveljournal.domain.member.entity.Member;
-import com.traveljournal.domain.member.repository.BlockRepository;
+import com.traveljournal.domain.block.repository.BlockRepository;
 import com.traveljournal.domain.search.dto.MemberSearchResponse;
 import com.traveljournal.domain.search.repository.MemberSearchRepository;
 import lombok.RequiredArgsConstructor;
@@ -10,8 +10,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
